Update dataset.py
<Introduction>
To begin with, it is important to note that this project comes with assumptions around size of audio data used for training, testing, and validation. However, code does not validate these assumptions (at least in few spots I've found). Hence this change is to exist program when there is insufficient data present.

<Changes>
This change is for __music_feed_epoch() only for now. 
This change should also be applied to the other *_feed_epoch() (such as blizzard), but I don't have the audio data to test, so won't make the necessary changes.

<Insight>
Without this change, if (for example) the origianl audio data is less than about 2 hours, the size of the test and verification data (.npy files) will be too small (less than a single batch size, where the min-command line option for batch size is 64, which is 64 x 8 seconds of audio at 16000 samples/seconds), but no error will be detected and traning run will continue to run without test and verification, the tale tale signs being terminal I/O such as this (notice 0, 'nan' and +38 values):

Validation! Done!
epoch:40	total iters:5001	wall clock time:1.48h
>>> Lowest valid cost:3.40282346639e+38	 Corresponding test cost:3.40282346639e+38
	train cost:2.4836	total time:0.35h	per iter:0.251s
	valid cost:nan	total time:0.00h
	test  cost:0.0000	total time:0.00h
Saving params! Done!
Train info saved! And plotted!